type fix

+++ src/AOloopControl_DM/AOloopControl_DM_runtimecomp.c
@@ -176,7 +176,7 @@ int AOloopControl_DM_CombineChannels(
 		float maxvolt
 		)
 {
-    long naxis = 2;
+    uint8_t naxis = 2;
     uint32_t *size;
     long ch;
     char name[200];
@@ -413,7 +413,7 @@ int AOloopControl_DM_CombineChannels(
         printf("vOK = %d\n", vOK);
         if(vOK==0)
         {
-			printf("CREATING stream %s  %d axis, size = %ld x %ld\n", dmdispcombconf[DMindex].voltname, naxis, size[0], size[1]);
+			printf("CREATING stream %s  %d axis, size = %u x %u\n", dmdispcombconf[DMindex].voltname, naxis, size[0], size[1]);
             dmdispcombconf[DMindex].IDvolt = create_image_ID(dmdispcombconf[DMindex].voltname, naxis, size, _DATATYPE_UINT16, 1, 10);
             COREMOD_MEMORY_image_set_createsem(dmdispcombconf[DMindex].voltname, 10);
          }
